Use a single FilterByText component
