+++ client/omnichannel/departments/DepartmentsPage.js
@@ -1,27 +1,11 @@
-import React, { useState, useEffect } from 'react';
-import { TextInput, Button, Box, Icon } from '@rocket.chat/fuselage';
+import React from 'react';
+import { Button, Icon } from '@rocket.chat/fuselage';
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
 
 import Page from '../../components/basic/Page';
-import { useTranslation } from '../../contexts/TranslationContext';
+import FilterByText from '../../components/FilterByText';
 import GenericTable from '../../components/GenericTable';
 import { useRoute } from '../../contexts/RouterContext';
-
-
-const FilterByText = ({ setFilter, ...props }) => {
-	const t = useTranslation();
-	const [text, setText] = useState('');
-
-	const handleChange = useMutableCallback((event) => setText(event.currentTarget.value));
-	const onSubmit = useMutableCallback((e) => e.preventDefault());
-
-	useEffect(() => {
-		setFilter({ text });
-	}, [setFilter, text]);
-	return <Box mb='x16' is='form' onSubmit={onSubmit} display='flex' flexDirection='column' {...props}>
-		<TextInput flexShrink={0} placeholder={t('Search')} addon={<Icon name='magnifier' size='x20'/>} onChange={handleChange} value={text} />
-	</Box>;
-};
 
 function DepartmentsPage({
 	data,
@@ -52,7 +36,7 @@ function DepartmentsPage({
 					total={data && data.total}
 					setParams={setParams}
 					params={params}
-					renderFilter={({ onChange, ...props }) => <FilterByText setFilter={onChange} {...props} />}
+					renderFilter={({ onChange, ...props }) => <FilterByText onChange={onChange} {...props} />}
 				/>
 			</Page.Content>
 		</Page>
